// src/types/api/service/models/ReporterDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReporterDto
 */
export interface ReporterDto {
    /**
     * mapped.pid, 매핑된 언론인 Id
     * @type {number}
     * @memberof ReporterDto
     */
    pid?: number;
    /**
     * 저자 이름
     * @type {string}
     * @memberof ReporterDto
     */
    name?: string;
    /**
     * 저자 부서
     * @type {string}
     * @memberof ReporterDto
     */
    department?: string;
    /**
     * 직책
     * @type {string}
     * @memberof ReporterDto
     */
    position?: string;
    /**
     * 언론인 DB로 Link 사용 가능 여부, 언론인 DB visible 값
     * @type {boolean}
     * @memberof ReporterDto
     */
    flagLink?: boolean;
}

/**
 * Check if a given object implements the ReporterDto interface.
 */
export function instanceOfReporterDto(value: object): value is ReporterDto {
    return true;
}

export function ReporterDtoFromJSON(json: any): ReporterDto {
    return ReporterDtoFromJSONTyped(json, false);
}

export function ReporterDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReporterDto {
    if (json == null) {
        return json;
    }
    return {
        
        'pid': json['pid'] == null ? undefined : json['pid'],
        'name': json['name'] == null ? undefined : json['name'],
        'department': json['department'] == null ? undefined : json['department'],
        'position': json['position'] == null ? undefined : json['position'],
        'flagLink': json['flagLink'] == null ? undefined : json['flagLink'],
    };
}

export function ReporterDtoToJSON(value?: ReporterDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'pid': value['pid'],
        'name': value['name'],
        'department': value['department'],
        'position': value['position'],
        'flagLink': value['flagLink'],
    };
}